Добавил проверку прав доступа пользователя

+++ compose/profiles/profiles/services/profiles/service.py
@@ -22,6 +22,11 @@ from .repository import (
     ProfileRepository,
     ProfileRepositoryDep,
 )
+from ..auth import (
+    AbstractPermissionService,
+    PermissionServiceDep,
+    AbstractPermissionChecker,
+)
 
 
 class AbstractProfileService(abc.ABC):
@@ -46,11 +51,18 @@ class AbstractProfileService(abc.ABC):
 
 class ProfileService(AbstractProfileService):
     repository: ProfileRepository
+    permission_checker: AbstractPermissionChecker
 
-    def __init__(self, *, repository: ProfileRepository) -> None:
+    def __init__(self,
+                 *,
+                 repository: ProfileRepository,
+                 permission_service: AbstractPermissionService) -> None:
         self.repository = repository
+        self.permission_checker = permission_service.get_permission_checker()
 
     async def get(self, *, user_id: uuid.UUID) -> ReadProfileResponse:
+        await self.permission_checker.check_read_permission(user_id=user_id)
+
         profile = await self.repository.get(user_id=user_id)
 
         if profile is None:
@@ -62,6 +74,8 @@ class ProfileService(AbstractProfileService):
                      *,
                      user_id: uuid.UUID,
                      profile_create: ProfileCreate) -> ReadProfileResponse:
+        await self.permission_checker.check_create_permission(user_id=user_id)
+
         try:
             profile = await self.repository.create(
                 user_id=user_id,
@@ -77,6 +91,8 @@ class ProfileService(AbstractProfileService):
                      *,
                      user_id: uuid.UUID,
                      profile_update: ProfileUpdate) -> ReadProfileResponse:
+        await self.permission_checker.check_update_permission(user_id=user_id)
+
         try:
             rows_count = await self.repository.update(
                 user_id=user_id,
@@ -92,6 +108,8 @@ class ProfileService(AbstractProfileService):
         return await self.get(user_id=user_id)
 
     async def delete(self, *, user_id: uuid.UUID) -> DeleteProfileResponse:
+        await self.permission_checker.check_delete_permission(user_id=user_id)
+
         rows_count = await self.repository.delete(user_id=user_id)
 
         if not rows_count:
@@ -100,8 +118,9 @@ class ProfileService(AbstractProfileService):
         return DeleteProfileResponse(user_id=user_id)
 
 
-async def get_profile_service(repository: ProfileRepositoryDep) -> AbstractProfileService:
-    return ProfileService(repository=repository)
+async def get_profile_service(repository: ProfileRepositoryDep,
+                              permission_service: PermissionServiceDep) -> AbstractProfileService:
+    return ProfileService(repository=repository, permission_service=permission_service)
 
 
 ProfileServiceDep = Annotated[AbstractProfileService, Depends(get_profile_service)]
